#25 Push Refactoring
