+++ JamNavigator/JamNavigator/PushNotificationHandler.swift
@@ -1,0 +1,48 @@
+//
+//  PushNotificationHandler.swift
+//  JamNavigator
+//
+//  Created by Manabu Tonosaki on 2021/12/16.
+//
+
+import Foundation
+import Firebase
+import UserNotifications
+
+
+extension AppDelegate : UNUserNotificationCenterDelegate {
+    
+    // ローカル通知を出す
+    func pushLocal(title: String, body: String, delaySeconds: Double = 0.0, sound: UNNotificationSound = .default) {
+        let content = UNMutableNotificationContent()
+        content.title = title
+        content.body = body
+        content.sound = sound
+
+        let intervalTrigger = UNTimeIntervalNotificationTrigger(timeInterval: delaySeconds, repeats: false)
+        let request = UNNotificationRequest(identifier: "localpush-\(title)-\(body)", content: content, trigger: intervalTrigger)
+        UNUserNotificationCenter.current().add(request, withCompletionHandler: nil) // 通知登録
+    }
+
+    // HACK:アプリがフォアグランドにいるときにPUSHされたイベント処理
+    func userNotificationCenter(_ center: UNUserNotificationCenter, willPresent notification: UNNotification, withCompletionHandler completionHandler: @escaping (UNNotificationPresentationOptions) -> Void) {
+
+        let userInfo = notification.request.content.userInfo
+        if let messageID = userInfo["gcm.message_id"] {
+            print("Message ID: \(messageID)")
+        }
+        print("======= PUSH1 \(userInfo)")
+        //completionHandler([])
+        completionHandler([ .badge, .sound, .banner ])
+   }
+
+    // HACK:ユーザーが通知バナーをタップした時に発火するイベント処理
+    func userNotificationCenter(_ center: UNUserNotificationCenter, didReceive response: UNNotificationResponse, withCompletionHandler completionHandler: @escaping () -> Void) {
+        let userInfo = response.notification.request.content.userInfo
+        if let messageID = userInfo["gcm.message_id"] {
+            print("Message ID: \(messageID)")
+        }
+        print("======= PUSH2 \(userInfo)")
+        completionHandler()
+    }
+}
